--- a/lib/grape/http/headers.rb
+++ b/lib/grape/http/headers.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'grape/util/lazy_object'
+
 module Grape
   module Http
     module Headers
@@ -26,6 +28,29 @@ module Grape
       HTTP_ACCEPT            = 'HTTP_ACCEPT'
 
       FORMAT                 = 'format'
+
+      HTTP_HEADERS = Grape::Util::LazyObject.new do
+        common_http_headers = %w[
+          Version
+          Host
+          Connection
+          Cache-Control
+          Dnt
+          Upgrade-Insecure-Requests
+          User-Agent
+          Sec-Fetch-Dest
+          Accept
+          Sec-Fetch-Site
+          Sec-Fetch-Mode
+          Sec-Fetch-User
+          Accept-Encoding
+          Accept-Language
+          Cookie
+        ].freeze
+        common_http_headers.each_with_object({}) do |header, response|
+          response["HTTP_#{header.upcase.tr('-', '_')}"] = header
+        end.freeze
+      end
 
       def self.find_supported_method(route_method)
         Grape::Http::Headers::SUPPORTED_METHODS.detect { |supported_method| supported_method.casecmp(route_method).zero? }