--- a/lib/grape/middleware/base.rb
+++ b/lib/grape/middleware/base.rb
@@ -74,11 +74,9 @@ module Grape
       end
 
       def mime_types
-        types_without_params = {}
-        content_types.each_pair do |k, v|
+        @mime_type ||= content_types.each_pair.with_object({}) do |(k, v), types_without_params|
           types_without_params[v.split(';').first] = k
         end
-        types_without_params
       end
 
       private