--- a/lib/grape/util/cache.rb
+++ b/lib/grape/util/cache.rb
@@ -1,0 +1,20 @@
+# frozen_String_literal: true
+
+require 'singleton'
+require 'forwardable'
+
+module Grape
+  module Util
+    class Cache
+      include Singleton
+
+      attr_reader :cache
+
+      class << self
+        extend Forwardable
+        def_delegators :cache, :[]
+        def_delegators :instance, :cache
+      end
+    end
+  end
+end