--- a/lib/grape/util/reverse_stackable_values.rb
+++ b/lib/grape/util/reverse_stackable_values.rb
@@ -9,7 +9,7 @@ module Grape
 
       def concat_values(inherited_value, new_value)
         [].tap do |value|
-          value.concat(new_value)
+          value.concat(new_value) if new_value
           value.concat(inherited_value)
         end
       end