--- a/lib/grape/util/stackable_values.rb
+++ b/lib/grape/util/stackable_values.rb
@@ -5,30 +5,19 @@ require_relative 'base_inheritable'
 module Grape
   module Util
     class StackableValues < BaseInheritable
-      attr_reader :frozen_values
-
-      def initialize(*_args)
-        super
-
-        @frozen_values = {}
-      end
-
       # Even if there is no value, an empty array will be returned.
       def [](name)
-        return @frozen_values[name] if @frozen_values.key? name
+        inherited_value = inherited_values[name]
+        new_value = new_values[name]
 
-        inherited_value = @inherited_values[name]
-        new_value = @new_values[name] || []
-
-        return new_value unless inherited_value
+        return new_value || [] unless inherited_value
 
         concat_values(inherited_value, new_value)
       end
 
       def []=(name, value)
-        raise if @frozen_values.key? name
-        @new_values[name] ||= []
-        @new_values[name].push value
+        new_values[name] ||= []
+        new_values[name].push value
       end
 
       def to_hash
@@ -37,16 +26,12 @@ module Grape
         end
       end
 
-      def freeze_value(key)
-        @frozen_values[key] = self[key].freeze
-      end
-
       protected
 
       def concat_values(inherited_value, new_value)
         [].tap do |value|
           value.concat(inherited_value)
-          value.concat(new_value)
+          value.concat(new_value) if new_value
         end
       end
     end