Hello,

I'm still working on improving string allocations and this PR follows my work.

- I added some lazy lookup tables at some places for a better string allocations.
- I added a lazy object which kind of a proxy for executing code when needed. It's mainly used when building headers for request. I also added a lazy object const that contains converted well known http headers.
- I fixed a small leak in stackable_values for empty arrays. It's now generating empty_array only when needed
- I replaced some %w by actual const

Here's some stats based on our app:
###
Overall Grape object allocations (development environment)
###
```log
# master

Total allocated: 47115631 bytes (437131 objects)
Total retained:  6323963 bytes (16355 objects)

# this PR
Total allocated: 41291815 bytes (300978 objects)
Total retained:  5841323 bytes (11641 objects)
```
### 
benchmark/simple.rb ruby 2.4.9
###
```log
# master
eproulx@Erics-Macbook grape % git checkout master
Switched to branch 'master'
eproulx@Erics-Macbook grape % ruby benchmark/simple.rb
Warming up --------------------------------------
              simple   522.000  i/100ms
Calculating -------------------------------------
              simple      5.420k (± 5.8%) i/s -     27.144k in   5.025995s
eproulx@Erics-Macbook grape % ruby benchmark/simple.rb
Warming up --------------------------------------
              simple   531.000  i/100ms
Calculating -------------------------------------
              simple      5.490k (± 4.7%) i/s -     27.612k in   5.041801s
eproulx@Erics-Macbook grape % git checkout lazy_lookup_lazy_object
Switched to branch 'lazy_lookup_lazy_object'
eproulx@Erics-Macbook grape % ruby benchmark/simple.rb            
Warming up --------------------------------------
              simple   605.000  i/100ms
Calculating -------------------------------------
              simple      5.871k (±10.0%) i/s -     29.040k in   5.000857s
eproulx@Erics-Macbook grape % ruby benchmark/simple.rb
Warming up --------------------------------------
              simple   606.000  i/100ms
Calculating -------------------------------------
              simple      6.409k (± 4.0%) i/s -     32.118k in   5.019708s
```

Thanks